Finding most common results in search queries, step 1
The results for each search query have been saved in a separate text file in a folder on my computer. This code creates an Excel file where the content of each text file in the folder is saved in an individual sheet in an Excel file. The sheet name is named after the text file that it represents.