Create STEP 4.2 Finding most common results in search queries
Use the Excel file that was created in STEP 4.1 to check for the most common results in the search queries. The result is an overview where for each entry, it is marked (with an 'x') in which of the search entries they appeared. The code is not without fault, so check under count._matches for those entries that are marked with 'FALSE'. They can be checked by using the EXCEL file created in STEP 4.1, to see under which search queries that particular paper ID appeared.